+++ unittests/data/string_traits.hpp
@@ -9,11 +9,14 @@ namespace string_traits{
 namespace yes{
     typedef std::string x1;
     typedef const std::string x2;
+    typedef std::string& x3;
+    typedef const std::string& x4;
 }
 
 namespace no{
     typedef int x1;
-    typedef std::string& x2;
+    typedef const int x2;
+    typedef const int& x3;
 }
 
 }
@@ -22,11 +25,14 @@ namespace wstring_traits{
 namespace yes{
     typedef std::wstring x1;
     typedef const std::wstring x2;
+    typedef std::wstring& x3;
+    typedef const std::wstring& x4;
 }
 
 namespace no{
     typedef int x1;
-    typedef std::wstring& x2;
+    typedef const int x2;
+    typedef const int& x3;
 }
 
 }